feat: Backend exchange voucher

+++ src/main/java/org/fatec/esportiva/repository/CepRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CepRepository extends JpaRepository<Cep, Integer> {
+public interface CepRepository extends JpaRepository<Cep, Long> {
      Optional<Cep> findByCep(String cep);
 }
