+++ src/main/java/org/fatec/esportiva/repository/ExchangeVoucherRepository.java
@@ -3,5 +3,5 @@ package org.fatec.esportiva.repository;
 import org.fatec.esportiva.entity.ExchangeVoucher;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExchangeVoucherRepository extends JpaRepository<ExchangeVoucher, Integer> {
+public interface ExchangeVoucherRepository extends JpaRepository<ExchangeVoucher, Long> {
 }
